refactor(claim-utils): move parseNumericValue to shared claim-utils.ts
Removes local copy from extract.ts and ingest-resource.ts — DRY violation
identified in paranoid review. Both files now import from the single
source of truth in crux/lib/claim-utils.ts.
